requirements.txt
einops
omegaconf
opencv-python
libigl
trimesh==4.2.0
pygltflib
pymeshlab==0.2
PyMCubes
xatlas
git+https://github.com/NVlabs/nvdiffrast/
scikit-learn
open_clip_torch
triton==2.1.0
rembg
gradio
tqdm
transformers==4.40.1
diffusers==0.19.3
ninja
imageio
imageio-ffmpeg
gradio-litmodel3d==0.0.1